requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlmodel==0.0.25
sqlalchemy==2.0.43
alembic==1.13.1
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic==2.11.9
pydantic-settings==2.11.0
email-validator==2.3.0
pyjwt[crypto]==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==1.2.0
ruff==0.3.5
black==24.2.0
pandas==2.2.2
plotly==5.21.0
streamlit==1.32.2
